Refactor electric field estimation into single function
The MATLAB version of FALCO contains separate functions for electric
field estimation using the full and compact models, which differ only in
the size of the return arrays and the model used to obtain the electric
field.  These two functions have been combined into a single function
that uses a keyword option to switch between the two cases.